requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3  # PostgreSQL을 사용하기 위해 필요
djangorestframework-simplejwt==5.1.0  # JWT 인증을 위한 패키지
dj-database-url==0.5.0
python-decouple
APScheduler==3.8.0